refactor: 맵 필터 기능(전체, 공공심야약국, 영업중) 유틸 함수로 분리
